refactor(extensions): extract sandbox srcdoc builder module
